timeline_view/requirements.txt
requests
beautifulsoup4
lxml
google-genai
google-cloud-storage
gcloud
airflow
